+++ src/parser/mod.rs
@@ -186,6 +186,7 @@ mod tests {
             ]
             .into_iter()
             .collect(),
+            levels: Default::default(),
         }
     }